backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart


transformers
torch
torchvision
pillow
protobuf
sentencepiece
safetensors
huggingface-hub
psycopg2-binary
